--- Project/src/main/java/com/codegym/project/controller/BannerController.java
@@ -1,17 +1,0 @@
-package com.codegym.project.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-@RequestMapping(value = "/banners")
-public class BannerController {
-    @GetMapping(value = "")
-    public ModelAndView listBanners(){
-        ModelAndView modelAndView=new ModelAndView("admin/Banner");
-        return modelAndView;
-    }
-
-}
