--- Project/src/main/java/com/codegym/project/controller/EmployeeController.java
@@ -1,16 +1,0 @@
-package com.codegym.project.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-@RequestMapping(value = "/employees")
-public class EmployeeController {
-    @GetMapping(value = "")
-    public ModelAndView listEmployees() {
-        ModelAndView modelAndView = new ModelAndView("admin/Employee");
-        return modelAndView;
-    }
-}
