--- Project/src/main/java/com/codegym/project/controller/OrderDetailController.java
@@ -1,12 +1,0 @@
-package com.codegym.project.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-public class OrderDetailController {
-//    @GetMapping(value = "")
-//    public ModelAndView listOrderDetail(){
-//        ModelAndView modelAndView=new ModelAndView("admin/Order");
-//        return modelAndView;
-//    }
-}
