--- Project/src/main/java/com/codegym/project/controller/ProductController.java
@@ -1,23 +1,0 @@
-package com.codegym.project.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-@RequestMapping(value = "")
-public class ProductController {
-
-    @GetMapping(value = "/products")
-    public ModelAndView listProducts(){
-        ModelAndView modelAndView=new ModelAndView("/admin/Product");
-        return modelAndView;
-    }
-
-    @GetMapping(value = "/productDeleted")
-    public ModelAndView listProductsDeleted(){
-        ModelAndView modelAndView= new ModelAndView("/admin/ProductIsDeleted");
-        return modelAndView;
-    }
-}
