--- Project/src/main/java/com/codegym/project/controller/TypeProductController.java
@@ -1,20 +1,0 @@
-package com.codegym.project.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-@Controller
-@RequestMapping(value = "")
-public class TypeProductController {
-    @GetMapping(value = "/typeProducts")
-    public ModelAndView listTypes(){
-        ModelAndView modelAndView=new ModelAndView("admin/ProductLine");
-        return modelAndView;
-    }
-    @GetMapping(value = "/typeProductsDeleted")
-    public ModelAndView listTypesDeleted(){
-        ModelAndView modelAndView=new ModelAndView("admin/ProductLineIsDelete");
-        return modelAndView;
-    }
-}
