--- Project/src/main/java/com/codegym/project/controller/VendorController.java
@@ -1,21 +1,0 @@
-package com.codegym.project.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-@Controller
-@RequestMapping(value = "")
-public class VendorController {
-    @GetMapping(value = "/vendors")
-    public ModelAndView listVendors(){
-        ModelAndView modelAndView=new ModelAndView("admin/Vendor");
-        return modelAndView;
-    }
-
-    @GetMapping(value = "/vendorsDeleted")
-    public ModelAndView listVendorsDeleted(){
-        ModelAndView modelAndView=new ModelAndView("admin/VendorIsDelete");
-        return modelAndView;
-    }
-}
